fix(dsp): preserve dtype in parallel band-group sum (#4125)
The band-group path summed worker results with np.sum(group_results, axis=0)
without the .astype(audio.dtype, copy=False) cast that the sibling
band-summation path applies (#3760). np.sum promotes to the widest worker
dtype, so a single float64 result (e.g. from scipy sosfiltfilt) promoted
float32 input to float64 — ~2x working memory and a dtype-contract violation.
Added the cast to match the sibling path. parallel-processor tests pass (25).

Closes #4125

+++ auralis/optimization/parallel_processor.py
@@ -401,8 +401,11 @@ class ParallelBandProcessor:
                             * (10 ** (band_gains[band_idx] / 20))
                         )
 
-        # Sum all group results
-        output: np.ndarray = np.sum(group_results, axis=0)
+        # Sum all group results. Cast back to the input dtype (#4125): np.sum
+        # promotes to the widest worker dtype, so a single float64 result (e.g.
+        # from scipy sosfiltfilt) would promote float32 input to float64 —
+        # matching the sibling band-summation path's #3760 fix.
+        output: np.ndarray = np.sum(group_results, axis=0).astype(audio.dtype, copy=False)
 
         return output
